+++ funciones/puestos/funciones_puestos.py
@@ -15,27 +15,31 @@ def PuestoDb():
         return render_template('ADMINISTRADOR/puestos/base_datos.html',titulo=titulo,puesto=puestosRecibidos)
     elif 'usuario-empleado' in session:
         return redirect('/INICIAR-SESION-EMPLEADO')
+
 def Formulario():
+    if 'usuario-administrador' in session:
         titulo="Nuevo operador"
         AuxiliaresDB=DB['auxiliar']
         AuxiliaresRecibidos=AuxiliaresDB.find()
         return render_template('ADMINISTRADOR/puestos/registrar.html', titulo=titulo,auxiliar=AuxiliaresRecibidos)
-       
+    elif 'usuario-empleado' in session:
+        return redirect('/INICIAR-SESION-EMPLEADO')
+
 def Nuevo():
     if 'usuario-administrador' in session:
         nombre=request.form['nombre']
         correo=request.form['correo']
         edad=request.form['edad']
         telefono=request.form['telefono']
-        tipo_puesto=request.form['tipo_puesto']
+        auxiliar=request.form['auxiliar']
         password=request.form['password']
         activo = False
         identificador=str(random.randint(0,2000))+correo
    
-        if nombre and correo and edad and telefono and tipo_puesto and password:
+        if nombre and correo and edad and telefono and auxiliar and password:
             key = generate_password_hash(password, method = 'sha256')
             puestos=DB['puestos']
-            puesto=Puesto(identificador,nombre,correo,edad,telefono,tipo_puesto,key,activo) 
+            puesto=Puesto(identificador,nombre,correo,edad,telefono,auxiliar,key,activo) 
             puestos.insert_one(puesto.datoPuestoJson())
             return redirect('/BASE-DATOS-OPERADOR')  
     
@@ -53,25 +57,34 @@ def Operaciones_Puesto():
         return redirect('/INICIAR-SESION-EMPLEADO')
 
 def Informacion_Puesto(key):
+    if 'usuario-administrador' in session:   
         titulo="Informacion puesto"
         puestos = DB['puestos']
         auxiliares = DB['auxiliar']
         auxiliar = auxiliares.find()    
         puestoRecibido=puestos.find_one({'identificador':key})
         return render_template('ADMINISTRADOR/puestos/informacion.html',titulo=titulo, puestos=puestoRecibido,aux=auxiliar) 
+    elif 'usuario-empleado' in session:
+        return redirect('/INICIAR-SESION-EMPLEADO')
 
 def Eliminar_Puesto(key):
-    puestos= DB['puestos']
-    puestos.delete_one({'identificador':key})
-    return redirect('/OPERACIONES-OPERADOR')
+    if 'usuario-administrador' in session:
+        puestos= DB['puestos']
+        puestos.delete_one({'identificador':key})
+        return redirect('/OPERACIONES-OPERADOR')
+    elif 'usuario-empleado'in session:
+        return redirect('/INICIAR-SESION-EMPLEADO')
 
 def Actualizar_puesto(key,campo):
-    puestos= DB['puestos']
-    dato=request.form['dato']
-    if dato:
-        puestos.update_one({'identificador':key},{'$set':{campo:dato}})
-    return Informacion_Puesto(key)
-    
+    if 'usuario-administrador' in session:
+        puestos= DB['puestos']
+        dato=request.form['dato']
+        if dato:
+            puestos.update_one({'identificador':key},{'$set':{campo:dato}})
+        return Informacion_Puesto(key)
+    elif 'usuario-empleado' in session:
+        return redirect('/INICIAR-SESION-EMPLEADO')
+
 def upload_operadores():
     file = request.files["upload_operadores"]
     operadores = DB['puestos']
